[CALCITE-2812] Add algebraic operators to allow expressing recursive queries (Ruben Quesada Lopez)
