+++ site/_docs/algebra.md
@@ -247,6 +247,53 @@ Similarly, to reference "DNAME", internal field #9 (8 + 1),
 write `builder.field(2, 1, "DNAME")`, `builder.field(2, "DEPT", "DNAME")`,
 or `builder.field(2, 1, 1)`.
 
+### Recursive Queries
+
+Warning: The current API is experimental and subject to change without notice.
+A SQL recursive query, e.g. this one that generates the sequence 1, 2, 3, ...10:
+
+{% highlight sql %}
+WITH RECURSIVE aux(i) AS (
+  VALUES (1)
+  UNION ALL
+  SELECT i+1 FROM aux WHERE i < 10
+)
+SELECT * FROM aux
+{% endhighlight %}
+
+can be generated using a scan on a TransientTable and a RepeatUnion:
+
+{% highlight java %}
+final RelNode node = builder
+  .values(new String[] { "i" }, 1)
+  .transientScan("aux")
+  .filter(
+      builder.call(
+          SqlStdOperatorTable.LESS_THAN,
+          builder.field(0),
+          builder.literal(10)))
+  .project(
+      builder.call(
+          SqlStdOperatorTable.PLUS,
+          builder.field(0),
+          builder.literal(1)))
+  .repeatUnion("aux", true)
+  .build();
+System.out.println(RelOptUtil.toString(node));
+{% endhighlight %}
+
+which produces:
+
+{% highlight text %}
+LogicalRepeatUnion(all=[true])
+  LogicalTableSpool(readType=[LAZY], writeType=[LAZY], tableName=[aux])
+    LogicalValues(tuples=[[{ 1 }]])
+  LogicalTableSpool(readType=[LAZY], writeType=[LAZY], tableName=[aux])
+    LogicalProject($f0=[+($0, 1)])
+      LogicalFilter(condition=[<($0, 10)])
+        LogicalTableScan(table=[[aux]])
+{% endhighlight %}
+
 ### API summary
 
 #### Relational operators
@@ -260,6 +307,7 @@ return the `RelBuilder`.
 |:------------------- |:-----------
 | `scan(tableName)` | Creates a [TableScan]({{ site.apiRoot }}/org/apache/calcite/rel/core/TableScan.html).
 | `functionScan(operator, n, expr...)`<br/>`functionScan(operator, n, exprList)` | Creates a [TableFunctionScan]({{ site.apiRoot }}/org/apache/calcite/rel/core/TableFunctionScan.html) of the `n` most recent relational expressions.
+| `transientScan(tableName [, rowType])` | Creates a [TableScan]({{ site.apiRoot }}/org/apache/calcite/rel/core/TableScan.html) on a [TransientTable]]({{ site.apiRoot }}/org/apache/calcite/schema/TransientTable.html) with the given type (if not specified, the most recent relational expression's type will be used).
 | `values(fieldNames, value...)`<br/>`values(rowType, tupleList)` | Creates a [Values]({{ site.apiRoot }}/org/apache/calcite/rel/core/Values.html).
 | `filter(expr...)`<br/>`filter(exprList)` | Creates a [Filter]({{ site.apiRoot }}/org/apache/calcite/rel/core/Filter.html) over the AND of the given predicates.
 | `project(expr...)`<br/>`project(exprList [, fieldNames])` | Creates a [Project]({{ site.apiRoot }}/org/apache/calcite/rel/core/Project.html). To override the default name, wrap expressions using `alias`, or specify the `fieldNames` argument.
@@ -279,6 +327,7 @@ return the `RelBuilder`.
 | `union(all [, n])` | Creates a [Union]({{ site.apiRoot }}/org/apache/calcite/rel/core/Union.html) of the `n` (default two) most recent relational expressions.
 | `intersect(all [, n])` | Creates an [Intersect]({{ site.apiRoot }}/org/apache/calcite/rel/core/Intersect.html) of the `n` (default two) most recent relational expressions.
 | `minus(all)` | Creates a [Minus]({{ site.apiRoot }}/org/apache/calcite/rel/core/Minus.html) of the two most recent relational expressions.
+| `repeatUnion(tableName, all [, n])` | Creates a [RepeatUnion]({{ site.apiRoot }}/org/apache/calcite/rel/core/RepeatUnion.html) associated to a [TransientTable]]({{ site.apiRoot }}/org/apache/calcite/schema/TransientTable.html) of the two most recent relational expressions, with `n` maximum number of iterations (default -1, i.e. no limit).
 | `snapshot(period)` | Creates a [Snapshot]({{ site.apiRoot }}/org/apache/calcite/rel/core/Snapshot.html) of the given snapshot period.
 | `match(pattern, strictStart,` `strictEnd, patterns, measures,` `after, subsets, allRows,` `partitionKeys, orderKeys,` `interval)` | Creates a [Match]({{ site.apiRoot }}/org/apache/calcite/rel/core/Match.html).
